major project/webapp-main/webapp/requirements.txt
flask
tensorflow
lime
Pillow
numpy
scikit-image
matplotlib
werkzeug